# design/radiology-openapi.yaml
openapi: 3.0.3
info:
  title: Radiology Plugin API
  description: |
    API specification for the Radiology Plugin for CARE.

    This plugin enables radiology workflows within CARE:
    - Upload DICOM files and retrieve studies from DCM4CHEE
    - Link DICOM studies to CARE service requests
    - Serve a DICOM modality worklist and receive study/MPPS status webhooks from the
      DICOM modality worklist server
    - Manage per-facility observation templates used to structure diagnostic report
      observations

    **Care Plugin Conventions:**
    - All `id` / `external_id` fields exposed via API are UUIDs (the table's `external_id`).
      The internal BigInt PK is never exposed.
    - All FK reference fields exposed via API are UUIDs (the referenced record's `external_id`).
    - `created_by` / `updated_by` are nested user objects, not bare IDs.

    **Authentication:**
    - Standard CARE endpoints use JWT Bearer auth (Authorization: Bearer <token>).
    - DICOM worklist and webhook endpoints use a static API key
      (Authorization: <CARE_RADIOLOGY_WEBHOOK_SECRET>).

  version: 1.0.0
  contact:
    name: Pranav Rustagi
    email: pranav.rutagi@egovernments.org
servers:
  - url: https://care.ohc.network/
    description: Dev server
  - url: https://care-k3s.digit.org/
    description: Egov server
...
security:
  - BearerAuth: []
paths:
...
  /api/care_radiology/dicom/authenticate/:
    get:
      tags:
        - DICOM Studies
      summary: JWT authentication verification
      description: |
        Dummy endpoint called by the nginx auth_request proxy to validate the JWT before
        forwarding DICOMweb requests to DCM4CHEE. Returns 200 if the token is valid.
      operationId: dicomAuthenticate
      responses:
        '200':
          description: Token is valid
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/care_radiology/dicom/worklist/:
    get:
      tags:
        - DICOM Studies
      summary: Get modality worklist
      description: |
        Returns active service requests formatted for a DICOM modality worklist.
        Authenticated with static API key (Authorization: <CARE_RADIOLOGY_WEBHOOK_SECRET>).
        Optionally filter by modality device name, date range, or facility.
      operationId: dicomWorklist
      security:
        - StaticApiKey: []
      parameters:
        - name: modality
          in: query
          description: Filter by modality device registered_name (case-insensitive)
          schema:
            type: string
            example: "CT"
        - name: from
          in: query
          description: Start date filter (YYYY-MM-DD or YYYY-MM-DD HH:MM:SS)
          schema:
            type: string
            example: "2026-06-01"
        - name: to
          in: query
          description: End date filter (YYYY-MM-DD or YYYY-MM-DD HH:MM:SS)
          schema:
            type: string
            example: "2026-06-12"
        - name: facility
          in: query
          description: Facility external_id (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Worklist results
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorklistEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/care_radiology/dicom/upload-dicom-external/:
    post:
      tags:
        - DICOM Studies
      summary: Upload DICOM file (external / static API key)
      description: |
        Upload a DICOM (.dcm) file for a patient using a static API key instead of
        user JWT auth. Intended for machine-to-machine use (e.g. modality worklist
        server, HIS integration). The file is forwarded to DCM4CHEE and a `DicomStudy`
        record is created or updated in CARE.

        Authenticated with static API key (Authorization: <CARE_RADIOLOGY_WEBHOOK_SECRET>).
        No user permission check is performed.
      operationId: dicomUploadExternal
      security:
        - StaticApiKey: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - patient_id
                - file
              properties:
                patient_id:
                  type: string
                  format: uuid
                  description: Patient external_id
                file:
                  type: string
                  format: binary
                  description: DICOM (.dcm) file
      responses:
        '201':
          description: DICOM file uploaded and study record created/updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "DICOM files uploaded to DCM4CHE successfully"
                  study_uid:
                    type: string
                    description: DICOM Study Instance UID
                    example: "1.2.840.10008.5.1.4.1.1.2.1234"
                  study:
                    $ref: '#/components/schemas/DicomStudyDetail'
        '400':
          description: No file provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "No file provided"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          description: DCM4CHEE upload failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Failed to upload to DCM4CHE"
                status_code: 503
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/care_radiology/dicom/upload/:
    post:
      tags:
        - DICOM Studies
      summary: Upload DICOM file
      description: |
        Upload a DICOM (.dcm) file for a patient. The file is forwarded to DCM4CHEE,
        the study UID is extracted from the response, and a `DicomStudy` record is
        created or updated in CARE. The study is then cached and returned.

        Requires `can_write_patient_obj` permission on the patient and
        `can_write_radiology_data` permission on the given facility.
      operationId: dicomUpload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - patient_id
                - facility_id
                - file
              properties:
                patient_id:
                  type: string
                  format: uuid
                  description: Patient external_id
                facility_id:
                  type: string
                  format: uuid
                  description: Facility external_id — used to authorize this upload against radiology data permissions
                file:
                  type: string
                  format: binary
                  description: DICOM (.dcm) file
      responses:
        '201':
          description: DICOM file uploaded and study record created/updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "DICOM files uploaded to Orthanc successfully"
                  study_uid:
                    type: string
                    description: DICOM Study Instance UID
                    example: "1.2.840.10008.5.1.4.1.1.2.1234"
                  study:
                    $ref: '#/components/schemas/DicomStudyDetail'
        '400':
          description: No file provided, or facility_id missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                no_file:
                  value:
                    error: "No file provided"
                no_facility:
                  value:
                    detail: "facility_id is required"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '502':
          description: DCM4CHEE upload failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Failed to upload to DCM4CHE"
                status_code: 503
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/care_radiology/dicom/link-service-request/:
    post:
      tags:
        - DICOM Studies
      summary: Link an uploaded study to a service request
      description: |
        Links an already-uploaded DICOM study to a CARE `ServiceRequest`. Called once
        all files for a study have finished uploading via `upload` / `upload-dicom-external`.

        Requires `can_write_service_request` permission on the service request, and
        `can_write_radiology_data` permission on the service request's facility.
      operationId: dicomLinkServiceRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - service_request_id
                - study_uid
              properties:
                service_request_id:
                  type: string
                  format: uuid
                  description: ServiceRequest external_id
                study_uid:
                  type: string
                  description: DICOM Study Instance UID
      responses:
        '200':
          description: Service request linked to study successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: "Service request linked to study successfully"
                  record:
                    type: object
                    properties:
                      external_id:
                        type: string
                        format: uuid
                        description: RadiologyServiceRequest external_id
                      data:
                        type: object
                        description: Raw payload stored on the link record
        '400':
          description: service_request_id and study_uid are required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "service_request_id and study_uid are required"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/care_radiology/dicom/studies/:
    get:
      tags:
        - DICOM Studies
      summary: Get DICOM studies for an encounter
      description: |
        Returns all DICOM studies linked to the given encounter, fetching live
        metadata (study date, description, modalities, series) from DCM4CHEE in parallel
        threads. Studies not found in DCM4CHEE are omitted from the response.

        Requires `can_view_encounter_obj` permission on the encounter, plus
        `can_read_radiology_data` permission on the encounter's facility.

        Note: a previous `patientId` query param was removed — `Patient` has no single
        owning facility to gate `can_read_radiology_data` against, so patient-scoped
        lookups are no longer supported on this endpoint.
      operationId: dicomGetStudies
      parameters:
        - name: encounterId
          in: query
          required: true
          description: Encounter external_id (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of DICOM studies with live DCM4CHEE metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DicomStudyDetail'
        '400':
          description: encounterId was not provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "encounterId is required"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/care_radiology/dicom/service-requests/:
    get:
      tags:
        - DICOM Studies
      summary: Get DICOM studies for a service request
      description: |
        Returns all DICOM studies linked to a given service request (via
        `RadiologyServiceRequest` join), including live DCM4CHEE metadata and
        the full service request detail.

        Requires `can_write_service_request` permission on the service request (a read-only
        endpoint gated by a write permission — pre-existing quirk), plus
        `can_read_radiology_data` permission on the service request's facility.
      operationId: dicomGetServiceRequests
      parameters:
        - name: serviceRequestId
          in: query
          required: true
          description: ServiceRequest external_id (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of service request + DICOM study pairs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceRequestStudyEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------- Webhooks ----------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Standard CARE JWT token
    StaticApiKey:
      type: apiKey
      in: header
      name: Authorization
      description: Static API key (CARE_RADIOLOGY_WEBHOOK_SECRET setting value)
...
  schemas:
...
    DicomSeries:
      type: object
      description: Summary of one DICOM series within a study, fetched live from DCM4CHEE
      properties:
        series_uid:
          type: string
          description: Series Instance UID
          example: "1.2.840.10008.5.1.4.1.1.2.456"
        series_number:
          type: array
          items:
            type: integer
        series_instance_count:
          type: array
          items:
            type: integer
          description: Number of instances in this series
        series_description:
          type: array
          items:
            type: string
        series_modality:
          type: array
          items:
            type: string
          example: ["CT"]
    DicomStudyDetail:
      type: object
      description: DICOM study metadata fetched from DCM4CHEE, combined with CARE record fields
      properties:
        external_id:
          type: string
          format: uuid
          description: CARE DicomStudy external_id
        study_uid:
          type: string
          description: DICOM Study Instance UID
          example: "1.2.840.10008.5.1.4.1.1.2.1234567890"
        study_date:
          type: string
          format: date-time
          nullable: true
          description: Study date/time in ISO 8601 format
          example: "2026-06-01T10:30:00"
        study_description:
          type: string
          nullable: true
          example: "CHEST PA"
        study_modalities:
          type: array
          items:
            type: string
          description: Modality codes present in the study
          example: ["CT"]
        study_series:
          type: array
          items:
            $ref: '#/components/schemas/DicomSeries'
...
    ServiceRequestDetail:
      type: object
      description: |
        Service request as returned by `ServiceRequestReadSpec`.
        The `encounter` field is serialized via `EncounterRetrieveSpec`.
      properties:
        id:
          type: string
          format: uuid
          description: ServiceRequest external_id
        title:
          type: string
        status:
          type: string
          enum: [draft, active, on_hold, entered_in_error, ended, completed, revoked]
        intent:
          type: string
          enum: [proposal, plan, directive, order]
        priority:
          type: string
          enum: [routine, urgent, asap, stat]
        category:
          type: string
        do_not_perform:
          type: boolean
          nullable: true
        note:
          type: string
          nullable: true
        code:
          type: object
          description: ValueSet-bound coding for the procedure code
        body_site:
          type: object
          nullable: true
        occurance:
          type: string
          format: date-time
          nullable: true
        patient_instruction:
          type: string
          nullable: true
        created_date:
          type: string
          format: date-time
        modified_date:
          type: string
          format: date-time
        encounter:
          $ref: '#/components/schemas/EncounterDetail'
        tags:
          type: array
          items:
            type: object
        requester:
          type: object
          nullable: true
          description: Nested user object (UserSpec)
    ServiceRequestStudyEntry:
      type: object
      description: A single entry returned by the service-requests DICOM endpoint — pairs the full service request with its linked DICOM study metadata.
      properties:
        service_request:
          $ref: '#/components/schemas/ServiceRequestDetail'
        dicom_study:
          allOf:
            - $ref: '#/components/schemas/DicomStudyDetail'
          nullable: true
          description: Live DCM4CHEE study metadata; null if the study is not found in DCM4CHEE
    WorklistEntry:
      type: object
      description: A single modality worklist entry (active service request formatted for DICOM MWL)
      properties:
        service_request:
          type: object
          properties:
            id:
              type: integer
              description: Internal DB id (not exposed to clients; included for DICOM MWL use)
            external_id:
              type: string
              format: uuid
            name:
              type: string
              description: Name of the activity/procedure
            date:
              type: string
              format: date-time
        facility:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        patient:
          type: object
          properties:
            id:
              type: integer
              description: Internal DB id
            external_id:
              type: string
              format: uuid
            name:
              type: string
            address:
              type: string
              nullable: true
            phone_number:
              type: string
              nullable: true
            gender:
              type: string
              nullable: true
            age:
              type: integer
              nullable: true

    # ===================== Webhooks =====================
...
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
        details:
          type: object
          nullable: true
          description: Field-level validation errors
          example:
            title: ["This field is required"]
  responses:
...
    Unauthorized:
      description: Unauthorized — missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Authentication credentials were not provided"
    Forbidden:
      description: Forbidden — authenticated but not permitted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "You do not have permission to perform this action"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Resource not found"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error occurred"